% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{d.G}
\alias{d.G}
\title{Transformation density for EGPD}
\usage{
d.G(u, type = 1, prob = NA, kappa = NA, delta = NA, log = FALSE)
}
\arguments{
\item{u}{values in (0,1)}

\item{type}{integer 1-6 specifying the transformation type}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter}

\item{delta}{shape parameter (types 4-6)}

\item{log}{logical: return log-density?}
}
\value{
Transformation density values
}
\description{
Transformation density for EGPD
}
