% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgpd.R
\name{dbgpd}
\alias{dbgpd}
\title{Density of the blended generalized Pareto distribution}
\usage{
dbgpd(
  x,
  xi,
  pbulk = pbeta,
  dbulk = dbeta,
  qbulk = qbeta,
  bulk.args = list(shape1 = 2, shape2 = 5),
  pa = 0.8,
  pb = 0.99,
  c1 = 5,
  c2 = 5,
  log = FALSE
)
}
\arguments{
\item{x}{values.}

\item{xi}{GP tail shape parameter.}

\item{pbulk}{bulk CDF supported on \eqn{[0, 1]}.}

\item{dbulk}{bulk density supported on \eqn{[0, 1]}. This is used to
validate the blended construction, even for \code{pbgpd()} and
\code{qbgpd()}.}

\item{qbulk}{bulk quantile function supported on \eqn{[0, 1]}.}

\item{bulk.args}{named list of arguments passed to \code{pbulk},
\code{dbulk}, and \code{qbulk}. Defaults to a Beta bulk.}

\item{pa}{lower bulk quantile used to start blending.}

\item{pb}{upper bulk quantile used to end blending and recover the exact GP
tail.}

\item{c1, c2}{Beta weighting-function shape parameters.}

\item{log}{logical; should log-densities be returned?}
}
\value{
Density values.
}
\description{
Density of the blended generalized Pareto distribution
}
