% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comppareto.R
\name{dcomppareto}
\alias{dcomppareto}
\title{Density of the continuous composite Pareto distribution}
\usage{
dcomppareto(x, spec, alpha = 1, theta = 1, log = FALSE, ...)
}
\arguments{
\item{x}{vector of quantiles.}

\item{spec}{body distribution specification. Currently one of
\code{"lnorm"}, \code{"gamma"}, \code{"weibull"}, or \code{"exp"}.}

\item{alpha}{positive Pareto shape parameter.}

\item{theta}{positive splice point and Pareto scale parameter.}

\item{log}{logical; return log-density?}

\item{...}{body-distribution parameters for the chosen \code{spec}.}
}
\value{
Numeric vector of densities.
}
\description{
Density of the continuous composite Pareto distribution
}
