% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpegpd.R
\name{dcpegpd}
\alias{dcpegpd}
\title{Density (PMF) of Compound Poisson-EGPD}
\usage{
dcpegpd(
  x,
  lambda,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma,
  xi,
  type = 1,
  h = 0.2,
  K = NULL,
  log = FALSE
)
}
\arguments{
\item{x}{numeric vector of quantiles}

\item{lambda}{Poisson rate parameter (> 0)}

\item{prob}{G-transformation parameter (type 6 only)}

\item{kappa}{G-transformation parameter}

\item{delta}{G-transformation parameter}

\item{sigma}{GPD scale parameter (> 0)}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G-transformation type}

\item{h}{bin width for discretization (default 0.2)}

\item{K}{number of bins (default: ceiling(max(x)/h) + 50)}

\item{log}{logical; if TRUE, return log-probabilities}
}
\value{
numeric vector of (log-)probabilities
}
\description{
Computes the probability mass function of the Compound Poisson-EGPD
distribution using Panjer recursion on a discretized EGPD severity.
}
