% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ddiscegpd}
\alias{ddiscegpd}
\title{Density of the Discrete Extended GPD}
\usage{
ddiscegpd(x, prob = NA, kappa = NA, delta = NA, sigma = NA, xi = NA, type = 1)
}
\arguments{
\item{x}{non-negative integer values}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter for G transformation}

\item{delta}{shape parameter for G transformation (types 4-6)}

\item{sigma}{GPD scale parameter}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G type}
}
\value{
PMF values
}
\description{
Density of the Discrete Extended GPD
}
\references{
Ahmad, T. and Arshad, I. A. (2024).
New flexible versions of extended generalized Pareto model for count data.
\emph{arXiv preprint} arXiv:2409.18719.
\url{https://arxiv.org/abs/2409.18719}
}
