% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{degpd_density}
\alias{degpd_density}
\title{Density of the Extended GPD}
\usage{
degpd_density(
  x,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma = NA,
  xi = NA,
  type = 1,
  log = FALSE
)
}
\arguments{
\item{x}{values}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter for G transformation}

\item{delta}{shape parameter for G transformation (types 4-6)}

\item{sigma}{GPD scale parameter}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G type}

\item{log}{logical: return log-density?}
}
\value{
Density values
}
\description{
Density of the Extended GPD
}
