% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docvisits}
\alias{docvisits}
\title{Doctor Visits}
\format{
A data frame with 1812 rows and 1 variable:
\describe{
  \item{visits}{Number of doctor visits (non-negative integer).}
}
}
\source{
Winkelmann, R. (2004). Health care reform and the number of doctor
  visits -- an econometric analysis. \emph{Journal of Applied Econometrics},
  19(4), 455--472.
}
\usage{
docvisits
}
\description{
Number of doctor visits for 1812 individuals from the German
Socio-Economic Panel (GSOEP). Originally distributed in the \pkg{zic}
package.
}
\keyword{datasets}
