% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dzidiscegpd}
\alias{dzidiscegpd}
\title{Density of the Zero-Inflated Discrete Extended GPD}
\usage{
dzidiscegpd(
  x,
  pi = NA,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma = NA,
  xi = NA,
  type = 1
)
}
\arguments{
\item{x}{non-negative integer values}

\item{pi}{zero-inflation probability}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter for G transformation}

\item{delta}{shape parameter for G transformation (types 4-6)}

\item{sigma}{GPD scale parameter}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G type}
}
\value{
PMF values
}
\description{
Density of the Zero-Inflated Discrete Extended GPD
}
