% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dziegpd}
\alias{dziegpd}
\title{Density of the Zero-Inflated Extended GPD}
\usage{
dziegpd(
  x,
  pi = NA,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma = NA,
  xi = NA,
  type = 1,
  log = FALSE
)
}
\arguments{
\item{x}{non-negative values}

\item{pi}{zero-inflation probability}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter for G transformation}

\item{delta}{shape parameter for G transformation (types 4-6)}

\item{sigma}{GPD scale parameter}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G type}

\item{log}{logical: return log-density?}
}
\value{
Density values
}
\description{
Continuous zero-inflated EGPD density:
\eqn{h(x) = \pi I(x=0) + (1-\pi) f_{EGPD}(x)} for \eqn{x \ge 0}.
}
