% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamlss_families.R
\name{egpd_bamlss}
\alias{egpd_bamlss}
\title{bamlss family for continuous EGPD}
\usage{
egpd_bamlss(m = 1, ...)
}
\arguments{
\item{m}{integer 1--4 selecting the G transformation:
\describe{
  \item{1}{Power: \eqn{G(u) = u^\kappa}}
  \item{2}{Mixture: \eqn{G(u) = p u^{\kappa} + (1-p) u^{\delta}}}
  \item{3}{Incomplete beta}
  \item{4}{Power-beta}
}}

\item{...}{arguments passed to link specification}
}
\value{
An object of class \code{family.bamlss}
}
\description{
Creates a \code{family.bamlss} object for fitting continuous
Extended Generalized Pareto Distribution models with \code{bamlss()}.
}
