% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd_fit.R
\name{fitted.egpd}
\alias{fitted.egpd}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{egpd}(object, ...)
}
\arguments{
\item{object}{a fitted \code{egpd} object}

\item{...}{not used}
}
\value{
Fitted values extracted from the object
}
\description{
Extract Model Fitted Values
}
