% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd_fit.R
\name{logLik.egpd}
\alias{logLik.egpd}
\title{Log-likelihood from a fitted \code{egpd} object}
\usage{
\method{logLik}{egpd}(object, ...)
}
\arguments{
\item{object}{a fitted \code{egpd} object}

\item{...}{not used}
}
\value{
A \code{logLik} object
}
\description{
Log-likelihood from a fitted \code{egpd} object
}
