% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd_fit.R
\name{nobs.egpd}
\alias{nobs.egpd}
\title{Number of observations}
\usage{
\method{nobs}{egpd}(object, ...)
}
\arguments{
\item{object}{an egpd object}

\item{...}{not used}
}
\value{
integer
}
\description{
Number of observations
}
