% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw_offenses}
\alias{nsw_offenses}
\title{Gaming and Betting Offenses in New South Wales}
\format{
A data frame with 342 rows and 1 variable:
\describe{
  \item{offenses}{Number of gaming and betting offenses (non-negative integer).}
}
}
\source{
Australian Bureau of Statistics, New South Wales crime data.
}
\usage{
nsw_offenses
}
\description{
Counts of gaming and betting offenses recorded at 342 locations in
New South Wales, Australia.
}
\keyword{datasets}
