% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{p.G}
\alias{p.G}
\title{Transformation CDF for EGPD}
\usage{
p.G(u, type = 1, prob, kappa, delta)
}
\arguments{
\item{u}{values in (0,1)}

\item{type}{integer 1-6 specifying the transformation type}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter}

\item{delta}{shape parameter (types 4-6)}
}
\value{
Transformed probabilities
}
\description{
Transformation CDF for EGPD
}
