% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comppareto.R
\name{pcomppareto}
\alias{pcomppareto}
\title{Distribution function of the continuous composite Pareto distribution}
\usage{
pcomppareto(
  q,
  spec,
  alpha = 1,
  theta = 1,
  lower.tail = TRUE,
  log.p = FALSE,
  ...
)
}
\arguments{
\item{q}{vector of quantiles.}

\item{spec}{body distribution specification. Currently one of
\code{"lnorm"}, \code{"gamma"}, \code{"weibull"}, or \code{"exp"}.}

\item{alpha}{positive Pareto shape parameter.}

\item{theta}{positive splice point and Pareto scale parameter.}

\item{lower.tail}{logical; if \code{FALSE}, return the upper tail.}

\item{log.p}{logical; return log-probabilities?}

\item{...}{body-distribution parameters for the chosen \code{spec}.}
}
\value{
Numeric vector of probabilities.
}
\description{
Distribution function of the continuous composite Pareto distribution
}
