% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpegpd.R
\name{pcpdegpd}
\alias{pcpdegpd}
\title{Distribution function of Compound Poisson-Discrete EGPD}
\usage{
pcpdegpd(
  q,
  lambda,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma,
  xi,
  type = 1,
  K = NULL
)
}
\arguments{
\item{q}{numeric vector of quantiles}

\item{lambda}{Poisson rate parameter (> 0)}

\item{prob}{G-transformation parameter (type 6 only)}

\item{kappa}{G-transformation parameter}

\item{delta}{G-transformation parameter}

\item{sigma}{GPD scale parameter (> 0)}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G-transformation type}

\item{K}{grid size for Panjer recursion (default: max(max(x), 100) + 50)}
}
\value{
numeric vector of cumulative probabilities
}
\description{
Distribution function of Compound Poisson-Discrete EGPD
}
\seealso{
\code{\link{dcpdegpd}}, \code{\link{qcpdegpd}},
  \code{\link{rcpdegpd}}
}
