% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpegpd.R
\name{pcpegpd}
\alias{pcpegpd}
\title{Distribution function of Compound Poisson-EGPD}
\usage{
pcpegpd(
  q,
  lambda,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma,
  xi,
  type = 1,
  h = 0.2,
  K = NULL
)
}
\arguments{
\item{q}{numeric vector of quantiles}

\item{lambda}{Poisson rate parameter (> 0)}

\item{prob}{G-transformation parameter (type 6 only)}

\item{kappa}{G-transformation parameter}

\item{delta}{G-transformation parameter}

\item{sigma}{GPD scale parameter (> 0)}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G-transformation type}

\item{h}{bin width for discretization (default 0.2)}

\item{K}{number of bins (default: ceiling(max(x)/h) + 50)}
}
\value{
numeric vector of cumulative probabilities
}
\description{
Distribution function of Compound Poisson-EGPD
}
