% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd_fit.R
\name{pinv}
\alias{pinv}
\alias{ginv.egpd}
\title{Moore-Penrose pseudo-inverse of a matrix}
\usage{
pinv(x, tol = -1)

ginv.egpd(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{a matrix}

\item{tol}{a scalar tolerance}
}
\value{
A matrix
}
\description{
Moore-Penrose pseudo-inverse of a matrix
}
