% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd_fit.R
\name{plot.egpd}
\alias{plot.egpd}
\title{Plot a fitted \code{egpd} object}
\usage{
\method{plot}{egpd}(x, onepage = TRUE, which = NULL, main, ask = !onepage, ...)
}
\arguments{
\item{x}{a fitted \code{egpd} object}

\item{onepage}{logical: should all plots be on one page? Defaults to TRUE}

\item{which}{a vector of integers identifying which smooths to plot}

\item{main}{a character string or vector of plot titles}

\item{ask}{logical: ask to show next plots?}

\item{...}{extra arguments to pass to \code{mgcv::plot.gam}}
}
\value{
Plots representing smooth terms
}
\description{
Plot a fitted \code{egpd} object
}
