% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.egpd}
\alias{predict.egpd}
\title{Predictions from a fitted \code{egpd} object}
\usage{
\method{predict}{egpd}(
  object,
  newdata,
  type = "link",
  prob = NULL,
  se.fit = FALSE,
  marginal = TRUE,
  trace = 0,
  ...
)
}
\arguments{
\item{object}{a fitted \code{egpd} object}

\item{newdata}{a data frame}

\item{type}{a character string: "link", "response", "lpmatrix", or "quantile"}

\item{prob}{a scalar or vector of probabilities for quantile estimation}

\item{se.fit}{logical: should standard errors be returned? Defaults to FALSE}

\item{marginal}{logical: should uncertainty integrate out smoothing parameter uncertainty? Defaults to TRUE}

\item{trace}{an integer controlling output verbosity}

\item{...}{unused}
}
\value{
A data frame, list, or design matrix depending on \code{type}
}
\description{
Predictions from a fitted \code{egpd} object
}
