% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd_fit.R
\name{print.egpd}
\alias{print.egpd}
\title{Print a fitted \code{egpd} object}
\usage{
\method{print}{egpd}(x, ...)
}
\arguments{
\item{x}{a fitted \code{egpd} object}

\item{...}{not used}
}
\value{
The call of the object (invisibly)
}
\description{
Print a fitted \code{egpd} object
}
