% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pziegpd}
\alias{pziegpd}
\title{CDF of the Zero-Inflated Extended GPD}
\usage{
pziegpd(
  q,
  pi = NA,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma = NA,
  xi = NA,
  type = 1
)
}
\arguments{
\item{q}{non-negative quantiles}

\item{pi}{zero-inflation probability}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter for G transformation}

\item{delta}{shape parameter for G transformation (types 4-6)}

\item{sigma}{GPD scale parameter}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G type}
}
\value{
CDF values
}
\description{
CDF of the Zero-Inflated Extended GPD
}
\references{
Abbas, A., Ahmad, T. and Ahmad, I. (2025).
Modeling zero-inflated precipitation extremes.
\emph{arXiv preprint} arXiv:2504.11058.
\url{https://arxiv.org/abs/2504.11058}
}
