% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{q.G}
\alias{q.G}
\title{Inverse transformation (quantile) for EGPD}
\usage{
q.G(u, type = 1, prob = NA, kappa = NA, delta = NA)
}
\arguments{
\item{u}{values in (0,1)}

\item{type}{integer 1-6}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter}

\item{delta}{shape parameter (types 4-6)}
}
\value{
Inverse-transformed values
}
\description{
Inverse transformation (quantile) for EGPD
}
