% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comppareto.R
\name{qcomppareto}
\alias{qcomppareto}
\title{Quantile function of the continuous composite Pareto distribution}
\usage{
qcomppareto(p, spec, alpha = 1, theta = 1, log.p = FALSE, ...)
}
\arguments{
\item{p}{vector of probabilities.}

\item{spec}{body distribution specification. Currently one of
\code{"lnorm"}, \code{"gamma"}, \code{"weibull"}, or \code{"exp"}.}

\item{alpha}{positive Pareto shape parameter.}

\item{theta}{positive splice point and Pareto scale parameter.}

\item{log.p}{logical; if \code{TRUE}, \code{p} is supplied on the log scale.}

\item{...}{body-distribution parameters for the chosen \code{spec}.}
}
\value{
Numeric vector of quantiles.
}
\description{
Quantile function of the continuous composite Pareto distribution
}
