% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpegpd.R
\name{qcpdegpd}
\alias{qcpdegpd}
\title{Quantile function of Compound Poisson-Discrete EGPD}
\usage{
qcpdegpd(
  p,
  lambda,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma,
  xi,
  type = 1,
  K = NULL
)
}
\arguments{
\item{p}{numeric vector of probabilities}

\item{lambda}{Poisson rate parameter (> 0)}

\item{prob}{G-transformation parameter (type 6 only)}

\item{kappa}{G-transformation parameter}

\item{delta}{G-transformation parameter}

\item{sigma}{GPD scale parameter (> 0)}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G-transformation type}

\item{K}{grid size for Panjer recursion (default: max(max(x), 100) + 50)}
}
\value{
numeric vector of quantiles (non-negative integers)
}
\description{
Quantile function of Compound Poisson-Discrete EGPD
}
\seealso{
\code{\link{dcpdegpd}}, \code{\link{pcpdegpd}},
  \code{\link{rcpdegpd}}
}
