% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{r.G}
\alias{r.G}
\title{Random generation from transformation G}
\usage{
r.G(
  n,
  prob = NA,
  kappa = NA,
  delta = NA,
  type = 1,
  unifsamp = NULL,
  direct = FALSE
)
}
\arguments{
\item{n}{number of samples}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter}

\item{delta}{shape parameter (types 4-6)}

\item{type}{integer 1-6}

\item{unifsamp}{optional uniform samples}

\item{direct}{logical for type 6}
}
\value{
Random samples
}
\description{
Random generation from transformation G
}
