% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comppareto.R
\name{rcomppareto}
\alias{rcomppareto}
\title{Random generation from the continuous composite Pareto distribution}
\usage{
rcomppareto(n, spec, alpha = 1, theta = 1, ...)
}
\arguments{
\item{n}{number of observations.}

\item{spec}{body distribution specification. Currently one of
\code{"lnorm"}, \code{"gamma"}, \code{"weibull"}, or \code{"exp"}.}

\item{alpha}{positive Pareto shape parameter.}

\item{theta}{positive splice point and Pareto scale parameter.}

\item{...}{body-distribution parameters for the chosen \code{spec}.}
}
\value{
Numeric vector of random draws.
}
\description{
Random generation from the continuous composite Pareto distribution
}
