% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpegpd.R
\name{rcpdegpd}
\alias{rcpdegpd}
\title{Random generation from Compound Poisson-Discrete EGPD}
\usage{
rcpdegpd(n, lambda, prob = NA, kappa = NA, delta = NA, sigma, xi, type = 1)
}
\arguments{
\item{n}{number of observations}

\item{lambda}{Poisson rate parameter (> 0)}

\item{prob}{G-transformation parameter (type 6 only)}

\item{kappa}{G-transformation parameter}

\item{delta}{G-transformation parameter}

\item{sigma}{GPD scale parameter (> 0)}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G-transformation type}
}
\value{
integer vector of length \code{n}
}
\description{
Direct simulation from the Compound Poisson-Discrete EGPD distribution:
draw \eqn{N \sim \mathrm{Poisson}(\lambda)}, then sum \eqn{N} i.i.d.
Discrete-EGPD draws.
}
\seealso{
\code{\link{dcpdegpd}}, \code{\link{pcpdegpd}}, \code{\link{qcpdegpd}}
}
