% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{regpd}
\alias{regpd}
\title{Random generation from the Extended GPD}
\usage{
regpd(
  n,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma = NA,
  xi = NA,
  type = 1,
  unifsamp = NULL
)
}
\arguments{
\item{n}{number of samples}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter for G transformation}

\item{delta}{shape parameter for G transformation (types 4-6)}

\item{sigma}{GPD scale parameter}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G type}

\item{unifsamp}{optional uniform samples}
}
\value{
Random samples
}
\description{
Random generation from the Extended GPD
}
