% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rzidiscegpd}
\alias{rzidiscegpd}
\title{Random generation from the Zero-Inflated Discrete Extended GPD}
\usage{
rzidiscegpd(
  n,
  pi = NA,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma = NA,
  xi = NA,
  type = 1,
  unifsamp = NULL
)
}
\arguments{
\item{n}{number of samples}

\item{pi}{zero-inflation probability}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter for G transformation}

\item{delta}{shape parameter for G transformation (types 4-6)}

\item{sigma}{GPD scale parameter}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G type}

\item{unifsamp}{optional uniform samples}
}
\value{
Random non-negative integer samples
}
\description{
Random generation from the Zero-Inflated Discrete Extended GPD
}
\references{
Ahmad, T. and Arshad, I. A. (2024).
New flexible versions of extended generalized Pareto model for count data.
\emph{arXiv preprint} arXiv:2409.18719.
\url{https://arxiv.org/abs/2409.18719}

Ahmad, T. and Hussain, A. (2025).
Flexible model for varying levels of zeros and outliers in count data.
\emph{arXiv preprint} arXiv:2510.27365.
\url{https://arxiv.org/abs/2510.27365}
}
