% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rziegpd}
\alias{rziegpd}
\title{Random generation from the Zero-Inflated Extended GPD}
\usage{
rziegpd(
  n,
  pi = NA,
  prob = NA,
  kappa = NA,
  delta = NA,
  sigma = NA,
  xi = NA,
  type = 1
)
}
\arguments{
\item{n}{number of samples}

\item{pi}{zero-inflation probability}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter for G transformation}

\item{delta}{shape parameter for G transformation (types 4-6)}

\item{sigma}{GPD scale parameter}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G type}
}
\value{
Random non-negative samples
}
\description{
Random generation from the Zero-Inflated Extended GPD
}
