% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd_fit.R
\name{seq_between}
\alias{seq_between}
\title{Generate a sequence between a range}
\usage{
seq_between(x, length = NULL)
}
\arguments{
\item{x}{a 2-vector}

\item{length}{an integer}
}
\value{
A vector
}
\description{
Generate a sequence between a range
}
