% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd_fit.R
\name{simulate.egpd}
\alias{simulate.egpd}
\title{Simulations from a fitted \code{egpd} object}
\usage{
\method{simulate}{egpd}(
  object,
  nsim = 1000,
  seed = NULL,
  newdata,
  type = "link",
  probs = NULL,
  threshold = 0,
  marginal = TRUE,
  ...
)
}
\arguments{
\item{object}{a fitted \code{egpd} object}

\item{nsim}{an integer giving the number of simulations}

\item{seed}{an integer giving the seed for simulations}

\item{newdata}{a data frame}

\item{type}{a character string: "link" or "response"}

\item{probs}{a scalar or vector of probabilities}

\item{threshold}{a scalar added to simulations}

\item{marginal}{logical: should uncertainty integrate out smoothing parameter uncertainty?}

\item{...}{additional arguments}
}
\value{
Simulations of parameters
}
\description{
Simulations from a fitted \code{egpd} object
}
