% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpd_fit.R
\name{summary.egpd}
\alias{summary.egpd}
\alias{print.summary.egpd}
\title{Summary method for a fitted \code{egpd} object}
\usage{
\method{summary}{egpd}(object, ...)

\method{print}{summary.egpd}(x, ...)
}
\arguments{
\item{object}{a fitted \code{egpd} object}

\item{...}{not used}

\item{x}{a \code{summary.egpd} object}
}
\value{
A \code{summary.egpd} object
}
\description{
Summary method for a fitted \code{egpd} object
}
