% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{vcov.egpd}
\alias{vcov.egpd}
\title{Variance-covariance matrix for egpd fits}
\usage{
\method{vcov}{egpd}(object, ...)
}
\arguments{
\item{object}{A fitted \code{egpd} object.}

\item{...}{Not used.}
}
\value{
The variance-covariance matrix \code{object$Vp} with row and column
  names matching the coefficient names.
}
\description{
Variance-covariance matrix for egpd fits
}
