% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bell_gamlss.R
\name{dZIBELL}
\alias{dZIBELL}
\alias{pZIBELL}
\alias{qZIBELL}
\alias{rZIBELL}
\alias{ZIBELL}
\title{gamlss family for the zero-inflated Bell distribution}
\usage{
dZIBELL(x, mu = 1, sigma = 0.1, log = FALSE)

pZIBELL(q, mu = 1, sigma = 0.1, lower.tail = TRUE, log.p = FALSE)

qZIBELL(
  p,
  mu = 1,
  sigma = 0.1,
  lower.tail = TRUE,
  log.p = FALSE,
  max.value = 1e+05
)

rZIBELL(n, mu = 1, sigma = 0.1)

ZIBELL(mu.link = "log", sigma.link = "logit")
}
\arguments{
\item{x, q}{vector of (non-negative integer) quantiles.}

\item{mu}{mean parameter (\eqn{>0}).}

\item{sigma}{zero-inflation probability \eqn{\in[0,1)}.}

\item{log, log.p, lower.tail, max.value}{as in \code{\link{bell}}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}

\item{mu.link}{link function for the mean parameter.}

\item{sigma.link}{link function for the zero-inflation parameter.}
}
\value{
\code{ZIBELL} returns a \code{gamlss.family} object.
}
\description{
\code{ZIBELL()} adds a logit-link zero-inflation probability (\code{sigma}) to
\code{\link{BELL}}. \code{dZIBELL}, \code{pZIBELL}, \code{qZIBELL} and
\code{rZIBELL} are the corresponding d/p/q/r functions, with \code{sigma} the
zero-inflation probability.
}
\seealso{
\code{\link{BELL}}; \code{\link{zibell}}.
}
