% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamlss_families.R
\name{zibell_bamlss}
\alias{zibell_bamlss}
\title{bamlss family for the zero-inflated Bell distribution}
\usage{
zibell_bamlss(...)
}
\arguments{
\item{...}{arguments passed to link specification}
}
\value{
An object of class \code{family.bamlss}
}
\description{
Adds a logit-link zero-inflation probability \code{pi} to
\code{\link{bell_bamlss}} (mapped to the \code{sigma} argument of the
underlying ZIBELL functions).
}
