% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamlss_families.R
\name{zidegpd_bamlss}
\alias{zidegpd_bamlss}
\title{bamlss Family for Zero-Inflated Discrete EGPD (ZIDEGPD)}
\usage{
zidegpd_bamlss(m = 1, ...)
}
\arguments{
\item{m}{integer 1--4 selecting the G transformation type (see
  \code{\link{egpd_bamlss}} for details)}

\item{...}{arguments passed to link specification}
}
\value{
An object of class \code{family.bamlss}
}
\description{
Creates a \code{family.bamlss} object for fitting zero-inflated discrete
Extended Generalized Pareto Distribution (ZIDEGPD) models with
\code{bamlss()}. Uses \code{\link{dzidiscegpd}},
\code{\link{pzidiscegpd}}, \code{\link{qzidiscegpd}}, and
\code{\link{rzidiscegpd}} internally.
}
\references{
Ahmad, T. and Arshad, I. A. (2024).
New flexible versions of extended generalized Pareto model for count data.
\emph{arXiv preprint} arXiv:2409.18719.
\url{https://arxiv.org/abs/2409.18719}

Ahmad, T. and Hussain, A. (2025).
Flexible model for varying levels of zeros and outliers in count data.
\emph{arXiv preprint} arXiv:2510.27365.
\url{https://arxiv.org/abs/2510.27365}
}
\seealso{
\code{\link{ZIDEGPD1}}, \code{\link{egpd_bamlss}},
\code{\link{degpd_bamlss}}, \code{\link{ziegpd_bamlss}},
\code{\link{dzidiscegpd}}
}
