% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpig_gamlss.R
\name{dZIGPIG}
\alias{dZIGPIG}
\alias{pZIGPIG}
\alias{qZIGPIG}
\alias{rZIGPIG}
\alias{ZIGPIG}
\title{gamlss family for the zero-inflated generalised Poisson-inverse Gaussian}
\usage{
dZIGPIG(x, mu = 1, sigma = 0.5, nu = 0.5, tau = 0.1, log = FALSE)

pZIGPIG(
  q,
  mu = 1,
  sigma = 0.5,
  nu = 0.5,
  tau = 0.1,
  lower.tail = TRUE,
  log.p = FALSE
)

qZIGPIG(
  p,
  mu = 1,
  sigma = 0.5,
  nu = 0.5,
  tau = 0.1,
  lower.tail = TRUE,
  log.p = FALSE,
  max.value = 1e+05
)

rZIGPIG(n, mu = 1, sigma = 0.5, nu = 0.5, tau = 0.1)

ZIGPIG(
  mu.link = "log",
  sigma.link = "logit",
  nu.link = "logit",
  tau.link = "logit"
)
}
\arguments{
\item{x, q}{vector of (non-negative integer) quantiles.}

\item{mu}{mean parameter (\eqn{>0}).}

\item{sigma}{tail exponent \eqn{a\in(0,1)}.}

\item{nu}{down-weighting parameter \eqn{c\in(0,1)}.}

\item{tau}{zero-inflation probability \eqn{\in[0,1)}.}

\item{log, log.p, lower.tail, max.value}{as in \code{\link{gpig}}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}

\item{mu.link, sigma.link, nu.link}{link functions for the parameters.}

\item{tau.link}{link function for the zero-inflation parameter.}
}
\value{
\code{ZIGPIG} returns a \code{gamlss.family} object.
}
\description{
\code{ZIGPIG()} adds a logit-link zero-inflation probability \code{tau} to
\code{\link{GPIG}}. \code{dZIGPIG}, \code{pZIGPIG}, \code{qZIGPIG} and
\code{rZIGPIG} are the corresponding d/p/q/r functions.
}
\seealso{
\code{\link{GPIG}}; \code{\link{zigpig}}.
}
