% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamlss_families.R
\name{zigpig_bamlss}
\alias{zigpig_bamlss}
\title{bamlss family for the zero-inflated generalised Poisson-inverse Gaussian}
\usage{
zigpig_bamlss(...)
}
\arguments{
\item{...}{arguments passed to link specification}
}
\value{
An object of class \code{family.bamlss}
}
\description{
Adds a logit-link zero-inflation probability \code{pi} to
\code{\link{gpig_bamlss}} (mapped to the \code{tau} argument of the
underlying ZIGPIG functions).
}
