% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamlss_families.R
\name{degpd_bamlss}
\alias{degpd_bamlss}
\title{bamlss Family for Discrete EGPD (DEGPD)}
\usage{
degpd_bamlss(m = 1, ...)
}
\arguments{
\item{m}{integer 1--4 selecting the G transformation type:
  \describe{
    \item{1}{Power: \eqn{G(u) = u^\kappa}}
    \item{2}{Mixture: \eqn{G(u) = p u^{\kappa} + (1-p) u^{\delta}}}
    \item{3}{Incomplete beta}
    \item{4}{Power-beta}
  }}

\item{...}{arguments passed to link specification}
}
\value{
An object of class \code{family.bamlss}
}
\description{
Creates a \code{family.bamlss} object for fitting discrete Extended
Generalized Pareto Distribution (DEGPD) models with \code{bamlss()}.
Uses \code{\link{ddiscegpd}}, \code{\link{pdiscegpd}},
\code{\link{qdiscegpd}}, and \code{\link{rdiscegpd}} internally.
}
\references{
Ahmad, T. and Arshad, I. A. (2024).
New flexible versions of extended generalized Pareto model for count data.
\emph{arXiv preprint} arXiv:2409.18719.
\url{https://arxiv.org/abs/2409.18719}
}
\seealso{
\code{\link{DEGPD1}}, \code{\link{egpd_bamlss}},
\code{\link{zidegpd_bamlss}}, \code{\link{ddiscegpd}}
}
