% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamlss_families.R
\name{egpd_bamlss}
\alias{egpd_bamlss}
\title{bamlss Family for Continuous EGPD}
\usage{
egpd_bamlss(m = 1, ...)
}
\arguments{
\item{m}{integer 1-4 specifying the G-transformation model}

\item{...}{additional arguments}
}
\value{
An object of class \code{family.bamlss}
}
\description{
bamlss Family for Continuous EGPD
}
\references{
Abbas, A., Ahmad, T. and Ahmad, I. (2025).
Modeling zero-inflated precipitation extremes.
\emph{arXiv preprint} arXiv:2504.11058.
\url{https://arxiv.org/abs/2504.11058}
}
\seealso{
\code{\link{EGPD1}}, \code{\link{degpd_bamlss}},
\code{\link{ziegpd_bamlss}}
}
