% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qdiscegpd}
\alias{qdiscegpd}
\title{Quantile function of the Discrete Extended GPD}
\usage{
qdiscegpd(p, prob = NA, kappa = NA, delta = NA, sigma = NA, xi = NA, type = 1)
}
\arguments{
\item{p}{probabilities}

\item{prob}{mixing probability (type 6)}

\item{kappa}{shape parameter for G transformation}

\item{delta}{shape parameter for G transformation (types 4-6)}

\item{sigma}{GPD scale parameter}

\item{xi}{GPD shape parameter}

\item{type}{integer 1-6 specifying G type}
}
\value{
Quantile values (non-negative integers)
}
\description{
Quantile function of the Discrete Extended GPD
}
