% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamlss_families.R
\name{ziegpd_bamlss}
\alias{ziegpd_bamlss}
\title{bamlss family for zero-inflated continuous EGPD}
\usage{
ziegpd_bamlss(m = 1, ...)
}
\arguments{
\item{m}{integer 1--4 selecting the G transformation (see
\code{\link{egpd_bamlss}} for details)}

\item{...}{arguments passed to link specification}
}
\value{
An object of class \code{family.bamlss}
}
\description{
Creates a \code{family.bamlss} object for fitting continuous
zero-inflated Extended Generalized Pareto Distribution models
with \code{bamlss()}.
}
